+++ src/Glitch/Context.php
@@ -30,7 +30,7 @@ use Throwable;
 
 class Context implements LoggerAwareInterface
 {
-    public const VERSION = 'v0.18.14';
+    public const VERSION = 'v0.18.15';
 
     protected float $startTime;
     protected string $runMode = 'development';